+++ internal/e2e/high_byte_literal_test.go
@@ -16,7 +16,7 @@ function main(): i32 {
     var s: string = "\x00\x41\x7f\x80\xc8\xef\xfe\xff";
     var out: string = s.len().to_string() + ":";
     var i: i32 = 0;
-    while (i < s.len()) { out = out + s[i].to_string() + " "; i = i + 1; }
+    while (i < s.len()) { out = out + (s[i] as i32).to_string() + " "; i = i + 1; }
     write(out + "\n");
     return 0;
 }
